* Extract AbstractArtifactAstTestCase
