+++ modules/extensions-spring-support/src/test/java/org/mule/test/module/extension/data/sample/AbstractSampleDataTestCase.java
@@ -6,26 +6,19 @@
  */
 package org.mule.test.module.extension.data.sample;
 
-import static org.mule.sdk.api.stereotype.MuleStereotypes.CONFIG;
 import static org.mule.runtime.api.component.TypedComponentIdentifier.ComponentType.OPERATION;
 import static org.mule.runtime.api.component.TypedComponentIdentifier.ComponentType.SOURCE;
 import static org.mule.runtime.api.metadata.MediaType.APPLICATION_JSON;
 import static org.mule.runtime.api.metadata.MediaType.APPLICATION_XML;
-import static org.mule.runtime.ast.api.ArtifactType.APPLICATION;
 import static org.mule.runtime.ast.api.util.MuleAstUtils.createComponentParameterizationFromComponentAst;
-import static org.mule.runtime.ast.api.util.MuleAstUtils.emptyArtifact;
 import static org.mule.runtime.core.api.extension.provider.MuleExtensionModelProvider.getExtensionModel;
 import static org.mule.runtime.core.api.lifecycle.LifecycleUtils.initialiseIfNeeded;
-import static org.mule.runtime.module.extension.internal.loader.java.AbstractJavaExtensionModelLoader.TYPE_PROPERTY_NAME;
-import static org.mule.runtime.module.extension.internal.loader.java.AbstractJavaExtensionModelLoader.VERSION;
 
-import static java.util.stream.Collectors.toMap;
 import static java.lang.System.lineSeparator;
-import static java.lang.Thread.currentThread;
-import static java.util.Collections.emptyMap;
 import static java.util.Collections.emptySet;
 import static java.util.Optional.empty;
 import static java.util.Optional.of;
+import static java.util.stream.Collectors.toMap;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
@@ -33,41 +26,31 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.StringContains.containsString;
 import static org.junit.Assert.fail;
 
-import org.mule.functional.junit4.CachingAstXmlParser;
-import org.mule.runtime.api.component.TypedComponentIdentifier.ComponentType;
-import org.mule.runtime.api.dsl.DslResolvingContext;
+import org.mule.functional.junit4.AbstractArtifactAstTestCase;
 import org.mule.runtime.api.lifecycle.InitialisationException;
 import org.mule.runtime.api.message.Message;
 import org.mule.runtime.api.meta.model.ExtensionModel;
 import org.mule.runtime.api.meta.model.config.ConfigurationModel;
 import org.mule.runtime.api.parameterization.ComponentParameterization;
 import org.mule.runtime.api.sampledata.SampleDataResult;
-import org.mule.runtime.ast.api.ArtifactAst;
-import org.mule.runtime.ast.api.ComponentAst;
-import org.mule.runtime.extension.api.loader.ExtensionModelLoader;
 import org.mule.runtime.extension.api.runtime.config.ConfigurationProvider;
 import org.mule.runtime.module.extension.api.runtime.config.ConfigurationProviderFactory;
 import org.mule.runtime.module.extension.api.runtime.config.ExtensionDesignTimeResolversFactory;
-import org.mule.runtime.module.extension.internal.loader.java.DefaultJavaExtensionModelLoader;
 import org.mule.runtime.module.extension.internal.runtime.config.DefaultConfigurationProviderFactory;
 import org.mule.runtime.module.extension.internal.runtime.config.DefaultExtensionDesignTimeResolversFactory;
 import org.mule.sdk.api.data.sample.SampleDataException;
-import org.mule.tck.junit4.AbstractMuleContextTestCase;
 import org.mule.test.data.sample.extension.SampleDataExtension;
 
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Map.Entry;
+import java.util.Optional;
 import java.util.Set;
 
 import org.junit.Before;
 
-public abstract class AbstractSampleDataTestCase extends AbstractMuleContextTestCase {
-
-  private static final CachingAstXmlParser AST_PARSER =
-      new CachingAstXmlParser(true, false, emptyMap(), APPLICATION, emptyArtifact());
+public abstract class AbstractSampleDataTestCase extends AbstractArtifactAstTestCase {
 
   protected static final String EXPECTED_PAYLOAD = "my payload";
   protected static final String EXPECTED_ATTRIBUTES = "my attributes";
@@ -80,18 +63,6 @@ public abstract class AbstractSampleDataTestCase extends AbstractMuleContextTest
   private ConfigurationProviderFactory configurationProviderFactory;
   private SampleDataExecutor sampleDataExecutor;
 
-  private ArtifactAst appAst;
-
-  @Before
-  public void loadAst() {
-    appAst = AST_PARSER.parse(this.getClass().getName(),
-                              getRequiredExtensions(),
-                              this.getClass().getClassLoader(),
-                              empty(),
-                              null,
-                              new String[] {getConfigFile()});
-  }
-
   @Before
   public void createExtensionDesignTimeResolversFactory() throws InitialisationException {
     extensionDesignTimeResolversFactory = new DefaultExtensionDesignTimeResolversFactory();
@@ -102,39 +73,13 @@ public abstract class AbstractSampleDataTestCase extends AbstractMuleContextTest
     sampleDataExecutor = new SampleDataExecutor(extensionDesignTimeResolversFactory);
   }
 
-  protected abstract String getConfigFile();
-
+  @Override
   protected Set<ExtensionModel> getRequiredExtensions() {
     final var extensions = new HashSet<ExtensionModel>();
     extensions.add(getExtensionModel());
     sampleDataExtension = loadExtension(SampleDataExtension.class, emptySet());
     extensions.add(sampleDataExtension);
     return extensions;
-  }
-
-  protected ExtensionModel loadExtension(Class extension, Set<ExtensionModel> deps) {
-    DefaultJavaExtensionModelLoader loader = new DefaultJavaExtensionModelLoader();
-    return loadExtensionWithLoader(extension, deps, loader);
-  }
-
-  protected ExtensionModel loadExtensionWithLoader(Class extension, Set<ExtensionModel> deps,
-                                                   ExtensionModelLoader extensionModelLoader) {
-    Map<String, Object> ctx = new HashMap<>();
-    ctx.put(TYPE_PROPERTY_NAME, extension.getName());
-    ctx.put(VERSION, "4.10.0");
-    ctx.putAll(getExtensionLoaderContextAdditionalParameters());
-    return extensionModelLoader.loadExtensionModel(currentThread().getContextClassLoader(), DslResolvingContext.getDefault(deps),
-                                                   ctx);
-  }
-
-  /**
-   * Subclasses can override this method so that extension models are generated with an extension loading context that contains
-   * the parameters returned by this method.
-   *
-   * @return a map with parameters to be added to the extension loader context.
-   */
-  protected Map<String, Object> getExtensionLoaderContextAdditionalParameters() {
-    return emptyMap();
   }
 
   protected void assertMessage(SampleDataResult result, String payload, String attributes) {
@@ -189,24 +134,6 @@ public abstract class AbstractSampleDataTestCase extends AbstractMuleContextTest
                        configurationProvider);
   }
 
-  private ComponentAst getFlowComponent(String flowName, ComponentType componentType) {
-    return appAst.topLevelComponentsStream()
-        .filter(f -> componentIdEquals(f, flowName))
-        .flatMap(flowAst -> flowAst.directChildrenStream()
-            .filter(comp -> componentType.equals(comp.getComponentType())))
-        .findFirst().orElseThrow();
-  }
-
-  private Optional<String> configNameFromComponent(ComponentAst comp) {
-    return comp.getParameters()
-        .stream()
-        .filter(p -> p.getModel().getAllowedStereotypes()
-            .stream()
-            .anyMatch(as -> as.isAssignableTo(CONFIG)))
-        .map(p -> (String) p.getValue().getRight())
-        .findAny();
-  }
-
   protected SampleDataResult getSampleByComponentName(String componentName,
                                                       Map<String, Object> parameters,
                                                       String configName)
@@ -251,18 +178,6 @@ public abstract class AbstractSampleDataTestCase extends AbstractMuleContextTest
                                                                            null,
                                                                            SampleDataExecutor
                                                                                .getClassLoader(sampleDataExtension));
-  }
-
-  private ComponentAst getTopLevelComponent(String configName) {
-    return appAst.topLevelComponentsStream()
-        .filter(f -> componentIdEquals(f, configName))
-        .findFirst().orElseThrow();
-  }
-
-  private boolean componentIdEquals(ComponentAst component, String componentId) {
-    return component.getComponentId()
-        .map(componentId::equals)
-        .orElse(false);
   }
 
   protected Map<String, Object> getDefaultParameters() {
