+++ modules/spring-config/src/test/java/org/mule/runtime/config/internal/context/service/InjectParamsFromContextServiceProxyJavaxTestCase.java
@@ -185,7 +185,8 @@ public class InjectParamsFromContextServiceProxyJavaxTestCase extends AbstractMu
 
     var thrown = assertThrows(IllegalDependencyInjectionException.class,
                               () -> serviceProxy.augmented());
-    assertThat(thrown.getMessage(), is(format(MANY_CANDIDATES_ERROR_MSG_TEMPLATE, "augmented", "AmbiguousAugmentedMethodService")));
+    assertThat(thrown.getMessage(),
+               is(format(MANY_CANDIDATES_ERROR_MSG_TEMPLATE, "augmented", "AmbiguousAugmentedMethodService")));
 
     assertThat(augmentedParam, nullValue());
   }
